Move StatsSinceLabelFormatter to utils package
